pyproject.toml
[project]
name = "briefen-me"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "flask-login>=0.6.3",
    "flask-limiter>=3.5.0",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.3.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "psycopg2-binary>=2.9.9",
    "gunicorn>=21.2.0",
    "qrcode[pil]>=7.4.2",
    "flask-cors>=6.0.1",
    "pyjwt>=2.10.1",
    "user-agents>=2.2.0",
    "google-cloud-storage>=2.14.0",
    "flask-migrate>=4.0.5",
]

[dependency-groups]
dev = ["pytest>=8.0", "ruff>=0.9.0"]

[tool.ruff]
target-version = "py311"
line-length = 88
extend-exclude = ["migrations/"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort (import sorting)
    "UP",   # pyupgrade (modernize syntax for py311)
    "B",    # flake8-bugbear (common bugs)
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # function calls in argument defaults (Flask patterns)
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
